+++ src/App.js
@@ -3,9 +3,9 @@ import './App.css';
 import {
   Switch, Route, NavLink, Redirect,
 } from 'react-router-dom';
-import Job from './components/Job/Job';
+import Job from './container/Job';
 import Nav from './components/Nav/Nav';
-import Homepage from './container/Homepage';
+import JobList from './container/JobList';
 
 function App() {
   return (
@@ -18,7 +18,7 @@ function App() {
         </NavLink>
       </nav>
       <Switch>
-        <Route path="/" exact component={Homepage} />
+        <Route path="/" exact component={JobList} />
         <Route path="/job/:job" exact component={Job} />
         <Redirect to="/" />
       </Switch>
