--- src/container/Homepage.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import JobList from '../components/JobList/JobList';
-
-function Homepage() {
-  return (
-    <div>
-      <JobList />
-    </div>
-  );
-}
-
-export default Homepage;
